chore: update i18n
